requirements.txt
crewai
crewai_tools
streamlit
python-dotenv
requests

tavily-python

plotly


langchain_google_genai
langchain